+++ src/lib/actions/absence.ts
@@ -19,7 +19,7 @@ export async function createAbsenceAction(config: ProfileConfig, args?: BaseComm
   } else if (selected.absence_event.unit_type === "days_and_half_days") {
     await chooseHalfDayAbsence(accessToken, message, selected);
   } else {
-    throw new Error("Unknow type " + selected.absence_event.unit_type);
+    throw new Error("Unknown type " + selected.absence_event.unit_type);
   }
 
   term.green("✓ Absence created\n\n");